sci-contact-harvester/requirements.txt
python-dotenv
requests
beautifulsoup4
selenium
playwright
PyGithub
anthropic
serpapi
pandas
schedule
streamlit
lxml
